Always starts Layman with auth env.vars.
