+++ src/layman/map/rest_maps_authz_test.py
@@ -1,6 +1,8 @@
 import json
+import pytest
 from test import process, process_client
 liferay_mock = process.liferay_mock
+ensure_layman = process.ensure_layman
 
 
 def assert_operates_on(workspace, mapname, expected_layers, authz_headers):
@@ -18,14 +20,13 @@ def assert_operates_on(workspace, mapname, expected_layers, authz_headers):
             operates_on_value), f"Expected layers {expected_layers}, found {json.dumps(operates_on_value, indent=2)}"
 
 
-def test_map_with_unauthorized_layer(liferay_mock):
+@pytest.mark.usefixtures('ensure_layman', 'liferay_mock')
+def test_map_with_unauthorized_layer():
     username1 = 'test_map_with_unauthorized_layer_user1'
     layername1 = 'test_map_with_unauthorized_layer_layer1'
     mapname1 = 'test_map_with_unauthorized_layer_map1'
     username2 = 'test_map_with_unauthorized_layer_user2'
     layername2 = 'test_map_with_unauthorized_layer_layer2'
-
-    layman_process = process.start_layman(process.AUTHN_SETTINGS)
 
     user1_authz_headers = process_client.get_authz_headers(username1)
     user2_authz_headers = process_client.get_authz_headers(username2)
@@ -73,5 +74,3 @@ def test_map_with_unauthorized_layer(liferay_mock):
     process_client.delete_map(username1, mapname1, headers=user1_authz_headers)
     process_client.delete_layer(username1, layername1, headers=user1_authz_headers)
     process_client.delete_layer(username2, layername2, headers=user2_authz_headers)
-
-    process.stop_process(layman_process)
